requirements.txt
pandas
pillow
tqdm
bidict
pytesseract
alive-progress
